Move getPkgs method to an internal package and exported
