+++ tools/testdata/scenariog/foo/mgmt/2019-10-11/foo/stage/fooapi/interfaces.go
@@ -16,6 +16,7 @@ package fooapi
 
 import (
 	"context"
+
 	"github.com/Azure/azure-sdk-for-go/tools/testdata/scenariog/foo/mgmt/2019-10-11/foo"
 )
