+++ tools/testpkgs/scenrioe/mgmt/2019-10-11/foo/fooapi/interfaces.go
@@ -2,6 +2,7 @@ package fooapi
 
 import (
 	"context"
+
 	"github.com/Azure/azure-sdk-for-go/tools/testpkgs/scenrioa/foo"
 )
