+++ tools/versioner/cmd/root.go
@@ -16,10 +16,11 @@ package cmd
 
 import (
 	"fmt"
-	"github.com/spf13/cobra"
 	"os"
 	"path/filepath"
 	"strings"
+
+	"github.com/spf13/cobra"
 )
 
 var rootCmd = &cobra.Command{
