+++ tools/versioner/cmd/unstage.go
@@ -18,10 +18,6 @@ import (
 	"bufio"
 	"errors"
 	"fmt"
-	"github.com/Azure/azure-sdk-for-go/tools/apidiff/repo"
-	"github.com/Azure/azure-sdk-for-go/tools/internal/modinfo"
-	"github.com/Masterminds/semver"
-	"github.com/spf13/cobra"
 	"io"
 	"io/ioutil"
 	"os"
@@ -30,6 +26,11 @@ import (
 	"regexp"
 	"sort"
 	"strings"
+
+	"github.com/Azure/azure-sdk-for-go/tools/apidiff/repo"
+	"github.com/Azure/azure-sdk-for-go/tools/internal/modinfo"
+	"github.com/Masterminds/semver"
+	"github.com/spf13/cobra"
 )
 
 var unstageCmd = &cobra.Command{
